mod- 블록 abstract 정리1
변수, abstract함수 정리 : 물리구현부+필수일부

protected ControlArduino arduino;
// 블록 구현부
private GameObject UpObj = null;
private GameObject DownObj;
protected Collider2D[] colliders;
protected Collider2D UpCollider;
protected Collider2D DownCollider;
public GameObject ParentObj;
private Block sample;

void OnDrop(PointerEventData eventData)
void OnDrag(PointerEventData eventData)
void OnTriggerStay2D(Collider2D collision)
void SetUpColllider(bool s)
void SetDownColllider(bool s)
GameObject CheckParentObj()
bool CheckUpCollider()
bool CheckDownCollider()
