+++ ArduinoVer6.0_19_12_11/Assets/Assets/Arduino_Ver3.0/Assets/Assets/FinalUI/Assets/test/NewBlockBox/ServoBlock.cs
@@ -1,9 +1,8 @@
 ﻿using System.Collections;
 using UnityEngine;
-using UnityEngine.EventSystems;
 using UnityEngine.UI;
 
-public class ServoBlock : Block, IDragHandler, IDropHandler
+public class ServoBlock : Block
 {
     #region 변수
 
@@ -11,24 +10,14 @@ public class ServoBlock : Block, IDragHandler, IDropHandler
     //public Canvas canvas;
     //[SerializeField]
     //public Camera secondCamera;
-
-    private GameObject UpObj = null;
-    private GameObject DownObj;
-    private Collider2D[] colliders;
-    private Collider2D UpCollider;
-    private Collider2D DownCollider;
-    private ControlArduino arduino;
+    
     public int selectnum = 0;
     public Socket selectSecket;
     public bool UpConncet = false;
     public bool DownConnect = false;
-
-    public GameObject ParentObj;
+    
     public float value = 0f;
-
-    private Block sample;
-    private bool GetChild = false;
-
+    
     #endregion 변수
 
     private void Start()
@@ -117,32 +106,7 @@ public class ServoBlock : Block, IDragHandler, IDropHandler
         if (GetChild == false && s == true)
         { GameManager.closeArduino(); GameManager.SyncRun(); }
     }
-
-    public override void SetDownColllider(bool s)
-    {
-        if (DownCollider != null)
-        {
-            DownCollider.isTrigger = s;
-        }
-    }
-
-    public override void SetUPColllider(bool s)
-    {
-        if (UpCollider != null)
-        {
-            UpCollider.isTrigger = s;
-        }
-    }
-
-    public override bool CheckUoCollider()
-    {
-        return UpCollider.isTrigger;
-    }
-
-    public override bool CheckDownCollider()
-    {
-        return DownCollider.isTrigger;
-    }
+    
 
     public override IEnumerator GetCode(bool s)
     {
@@ -239,81 +203,8 @@ public class ServoBlock : Block, IDragHandler, IDropHandler
         }
     }
 
-    public override GameObject CheckParentObj()
-    {
-        return UpObj;
-    }
-
     #endregion 필수 구현 부분
 
-    #region 물리 구현 부분
-
-    public void OnTriggerStay2D(Collider2D collision)
-    {
-        if (collision.tag != "region")
-        {
-            if (collision == null)
-                return;
-
-            if (transform.position.y < collision.transform.position.y)//자기 위에 충돌할때
-            {
-                if (UpCollider.isTrigger == true)
-                {
-                    sample = BlockManager.instance.BlockIdentity(collision);
-                    if (sample != null)
-                    {
-                        if (sample.CheckDownCollider() == true)
-                        {
-                            transform.position = collision.transform.position + new Vector3(0, -51, 0);
-                            this.transform.SetParent(sample.transform);
-                            this.transform.SetAsFirstSibling();
-                            UpObj = collision.gameObject;
-                            UpCollider.isTrigger = false;
-                            sample.SetDownColllider(false);
-                        }
-                    }
-                }
-            }
-            else if (transform.position.y > collision.transform.position.y) // 자기 아랫부분에서 충돌할때
-            {
-                sample = BlockManager.instance.BlockIdentity(collision);
-                if (sample != null)
-                {
-                    if (sample.CheckParentObj() == this.gameObject)
-                    {
-                        DownObj = collision.gameObject;
-                    }
-                }
-            }
-        }
-    }
-
-    public void OnDrop(PointerEventData eventData)
-    {
-        this.GetComponent<Outline>().effectColor = new Color(255, 0, 0, 0);
-    }
-
-    public void OnDrag(PointerEventData eventData)
-    {
-        if (this.transform.parent != null)
-        {
-            if (UpObj != null)
-            {
-                Block block = BlockManager.instance.BlockIdentity(UpObj);
-                block.SetDownColllider(true);
-                UpCollider.isTrigger = true;
-
-                this.transform.SetParent(ParentObj.transform);
-            }
-        }
-
-        if (GameManager.RunBlock == true)
-            transform.position = Input.mousePosition; //secondCamera.ScreenToWorldPoint(screenpoint);
-
-        this.GetComponent<Outline>().effectColor = new Color(255, 0, 0, 255);
-    }
-
-    #endregion 물리 구현 부분
 
     #region 고유 구현 부분
