+++ ArduinoVer6.0_19_12_11/Assets/Assets/Arduino_Ver3.0/Assets/Assets/FinalUI/Assets/test/NewBlockBox/WaitBlock.cs
@@ -1,9 +1,8 @@
 ﻿using System.Collections;
 using UnityEngine;
-using UnityEngine.EventSystems;
 using UnityEngine.UI;
 
-public class WaitBlock : Block, IDragHandler, IDropHandler
+public class WaitBlock : Block
 {
     #region 변수
 
@@ -17,43 +16,14 @@ public class WaitBlock : Block, IDragHandler, IDropHandler
     //public Camera secondCamera;
     //private WaitForSeconds waitForSeconds;
 
-    private GameObject UpObj = null;
-    private GameObject DownObj;
-    private Collider2D[] colliders;
-    private Collider2D UpCollider;
-    private Collider2D DownCollider;
-    private ControlArduino arduino;
-    public GameObject ParentObj;
-
     public int TimeForWait = 0;
 
     private Block sample;
-    private bool GetChild = false;
-
     #endregion 변수
-
-    private void Start()
-    {
-        arduino = GameObject.FindWithTag("Arduino").GetComponent<ControlArduino>(); //아두이노 오브젝트 찾기
-        ParentObj =
-            GameObject.Find("PanelBlockCoding").
-            gameObject.transform.Find("CodingPanel").
-            gameObject.transform.Find("CodingMaskPanel"). //코딩마스크패널 찾기
-            gameObject;
-
-        this.transform.position = new Vector3(930, 421); //초기위치지정
-
-        colliders = this.GetComponents<Collider2D>();//위, 아래 충돌 지정
-        if (colliders != null)
-        {
-            DownCollider = colliders[0];
-
-            UpCollider = colliders[1];
-        }
-    }
+    
+    // public void Start()
 
     #region 필수 구현부분
-
     public override IEnumerator Run(float s)
     {
         Debug.Log("웨이트 개시");
@@ -98,45 +68,6 @@ public class WaitBlock : Block, IDragHandler, IDropHandler
         }
     }
 
-    public override IEnumerator GetBtCode(bool s)
-    {
-        GetChild = false;
-       
-        yield return new WaitForSeconds(0.3f);
-
-        Block block = BlockManager.instance.BlockIdentity(transform);
-        if (block != null)
-        {
-            StartCoroutine(block.GetBtCode(s));
-            GetChild = true;
-        }
-
-        if (GetChild == false && s == true)
-        {
-            GameManager.syncBTMergeCode();
-            GameManager.connectArduino.SetMeshMaterial(false);
-        }
-    }
-
-    public override IEnumerator GetSyncCode(bool s)
-    {
-        GetChild = false;
-      
-        yield return new WaitForSeconds(0.3f);
-
-        Block block = BlockManager.instance.BlockIdentity(transform);
-        if (block != null)
-        {
-            StartCoroutine(block.GetSyncCode(s));
-            GetChild = true;
-        }
-
-        if (GetChild == false && s == true)
-        {
-            GameManager.syncMergeCode();
-            GameManager.connectArduino.SetMeshMaterial(false);
-        }
-    }
 
     public override IEnumerator GetCode(bool s)
     {
@@ -160,107 +91,47 @@ public class WaitBlock : Block, IDragHandler, IDropHandler
         }
     } //코드 떼오기
 
-    public override void SetDownColllider(bool s)
+    public override IEnumerator GetSyncCode(bool s)
     {
-        if (DownCollider != null)
-        {
-            DownCollider.isTrigger = s;
-        }
-    } //업 다운 콜라이더 트리거 설정
+        GetChild = false;
+      
+        yield return new WaitForSeconds(0.3f);
 
-    public override void SetUPColllider(bool s)
-    {
-        if (UpCollider != null)
+        Block block = BlockManager.instance.BlockIdentity(transform);
+        if (block != null)
         {
-            UpCollider.isTrigger = s;
+            StartCoroutine(block.GetSyncCode(s));
+            GetChild = true;
+        }
+
+        if (GetChild == false && s == true)
+        {
+            GameManager.syncMergeCode();
+            GameManager.connectArduino.SetMeshMaterial(false);
         }
     }
 
-    public override bool CheckUoCollider()
+    public override IEnumerator GetBtCode(bool s)
     {
-        return UpCollider.isTrigger;
-    } //위 충돌 트리거상태 리턴
+        GetChild = false;
+       
+        yield return new WaitForSeconds(0.3f);
 
-    public override bool CheckDownCollider()
-    {
-        return DownCollider.isTrigger;
-    } //아래 충돌 트리거 상태 리턴
+        Block block = BlockManager.instance.BlockIdentity(transform);
+        if (block != null)
+        {
+            StartCoroutine(block.GetBtCode(s));
+            GetChild = true;
+        }
 
-    public override GameObject CheckParentObj()
-    {
-        return UpObj;
-    } //위에 있는 블록 리턴
-
+        if (GetChild == false && s == true)
+        {
+            GameManager.syncBTMergeCode();
+            GameManager.connectArduino.SetMeshMaterial(false);
+        }
+    }
     #endregion 필수 구현부분
-
-    #region 물리 구현 부분
-
-    public void OnDrop(PointerEventData eventData)
-    {
-        this.GetComponent<Outline>().effectColor = new Color(255, 0, 0, 0);
-    }
-
-    public void OnDrag(PointerEventData eventData)
-    {
-        if (this.transform.parent != null)
-        {
-            if (UpObj != null)
-            {
-                Block block = BlockManager.instance.BlockIdentity(UpObj);
-                block.SetDownColllider(true);
-                UpCollider.isTrigger = true;
-
-                this.transform.SetParent(ParentObj.transform);
-            }
-        }
-
-        if (GameManager.RunBlock == true)
-            transform.position = Input.mousePosition; //secondCamera.ScreenToWorldPoint(screenpoint);
-
-        this.GetComponent<Outline>().effectColor = new Color(255, 0, 0, 255);
-    }
-
-    public void OnTriggerStay2D(Collider2D collision)
-    {
-        if (collision.tag != "region")
-        {
-            if (collision == null)
-                return;
-
-            if (transform.position.y < collision.transform.position.y)//자기 위에 충돌할때
-            {
-                if (UpCollider.isTrigger == true)
-                {
-                    sample = BlockManager.instance.BlockIdentity(collision);
-                    if (sample != null)
-                    {
-                        if (sample.CheckDownCollider() == true)
-                        {
-                            transform.position = collision.transform.position + new Vector3(0, -51, 0);
-                            this.transform.SetParent(sample.transform);
-                            this.transform.SetAsFirstSibling();
-                            UpObj = collision.gameObject;
-                            UpCollider.isTrigger = false;
-                            sample.SetDownColllider(false);
-                        }
-                    }
-                }
-            }
-            else if (transform.position.y > collision.transform.position.y) // 자기 아랫부분에서 충돌할때
-            {
-                sample = BlockManager.instance.BlockIdentity(collision);
-                if (sample != null)
-                {
-                    if (sample.CheckParentObj() == this.gameObject)
-                    {
-                        DownObj = collision.gameObject;
-                    }
-                }
-            }
-        }
-    } //충돌시 붙히기 코드
-
-    #endregion 물리 구현 부분
+    
 
     #region 고유 구현 부분
