+++ ArduinoVer6.0_19_12_11/Assets/Assets/Arduino_Ver3.0/Assets/Assets/FinalUI/ifBlock.cs
@@ -17,15 +17,7 @@ public class ifBlock : Block, IDragHandler, IDropHandler
     //public bool selectRun = true;
     // public RectTransform barrect;
     //public Vector3 Firstlocation;
-
-    private GameObject UpObj;
-    private GameObject DownObj;
-    private Collider2D[] colliders;
-    private Collider2D UpCollider;
-
-    [SerializeField]
-    private Collider2D DownCollider;
-
+    
     public int ThirdSel = 0;
     private bool okvalue = false;
     public int FirstSel = 0;
@@ -38,10 +30,8 @@ public class ifBlock : Block, IDragHandler, IDropHandler
 
     public Block sample;
     public Vector2 FirstAnchoredPosition;
-
-    public GameObject ParentObj;
+    
     private Outline outline;
-    public bool GetChild = false;
 
     #endregion 변수
 
@@ -578,107 +568,8 @@ public class ifBlock : Block, IDragHandler, IDropHandler
         StartCoroutine(bar.GetBtCode(s));
     }
 
-    public override void SetDownColllider(bool s)
-    {
-        if (DownCollider != null)
-        {
-            DownCollider.isTrigger = s;
-        }
-    }
-
-    public override void SetUPColllider(bool s)
-    {
-        if (UpCollider != null)
-        {
-            UpCollider.isTrigger = s;
-        }
-    }
-
-    public override bool CheckUoCollider()
-    {
-        return UpCollider.isTrigger;
-    }
-
-    public override bool CheckDownCollider()
-    {
-        return DownCollider.isTrigger;
-    }
-
-    public override GameObject CheckParentObj()
-    {
-        return UpObj;
-    }
-
     #endregion 필수 구현부분
-
-    #region 물리 구현 부분
-
-    public void OnTriggerStay2D(Collider2D collision)
-    {
-        if (collision.tag != "region")
-        {
-            if (collision == null)
-                return;
-
-            if (transform.position.y < collision.transform.position.y)//자기 위에 충돌할때
-            {
-                if (UpCollider.isTrigger == true)
-                {
-                    sample = BlockManager.instance.BlockIdentity(collision);
-                    if (sample != null)
-                    {
-                        if (sample.CheckDownCollider() == true)
-                        {
-                            transform.position = collision.transform.position + new Vector3(0, -51, 0);
-                            this.transform.SetParent(sample.transform);
-                            this.transform.SetAsFirstSibling();
-                            UpObj = collision.gameObject;
-                            UpCollider.isTrigger = false;
-                            sample.SetDownColllider(false);
-                        }
-                    }
-                }
-            }
-            else if (transform.position.y > collision.transform.position.y) // 자기 아랫부분에서 충돌할때
-            {
-                sample = BlockManager.instance.BlockIdentity(collision);
-                if (sample != null)
-                {
-                    if (sample.CheckParentObj() == this.gameObject)
-                    {
-                        DownObj = collision.gameObject;
-                    }
-                }
-            }
-        }
-    }
-
-    public void OnDrop(PointerEventData eventData)
-    {
-        this.GetComponent<Outline>().effectColor = new Color(255, 0, 0, 0);
-    }
-
-    public void OnDrag(PointerEventData eventData)
-    {
-        if (this.transform.parent != null)
-        {
-            if (UpObj != null)
-            {
-                Block block = BlockManager.instance.BlockIdentity(UpObj);
-                block.SetDownColllider(true);
-                UpCollider.isTrigger = true;
-
-                this.transform.SetParent(ParentObj.transform);
-            }
-        }
-
-        if (GameManager.RunBlock == true)
-            transform.position = Input.mousePosition; //secondCamera.ScreenToWorldPoint(screenpoint);
-
-        this.GetComponent<Outline>().effectColor = new Color(255, 0, 0, 255);
-    }
-
-    #endregion 물리 구현 부분
+    
 
     #region 고유 구현 부분
